Reorganise indexable models
