+++ catalogue_graph/src/ingestor/models/indexable/record.py
