+++ catalogue_graph/src/ingestor/transformers/concepts_transformer.py
@@ -6,7 +6,7 @@ from ingestor.extractors.concepts.concepts_index_extractor import (
     ConceptsIndexExtractor,
     ExtractedRelatedConcepts,
 )
-from ingestor.models.indexable_concept import (
+from ingestor.models.indexable.concept import (
     ConceptDisplay,
     ConceptQuery,
     ConceptRelatedTo,
